requirements.txt
asgiref==3.8.1
Django==5.1.2
django-cors-headers==4.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
lxml==5.3.0
packaging==24.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-docx==1.1.2
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
